Refactor the different sections of the profile view into different components. #16
