+++ client/components/music-section.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
 import Parser from 'html-react-parser'
+import { Container } from 'reactstrap'
 
 export default class MusicSection extends Component {
   render() {
@@ -11,14 +12,14 @@ export default class MusicSection extends Component {
       }
     }
     return (
-      <div>
+      <Container className="clear-border border-right px-4 col-md-9 float-md-left">
         {links.map((link, index) => {
           return (
             <div key={ index }>
               { link }
             </div>)
         })}
-      </div>
+      </Container>
     )
   }
 }
