+++ client/components/profile-header.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react'
+
+export default class ProfileHeader extends Component {
+  render() {
+    const { displayName, firstName, lastName, city, country, imageUrl, artistType } = this.props.artist
+    return (
+      <div>
+        <div className="profile-header p-4">
+          <div className="profile-image rounded-circle float-md-left" style={{ backgroundImage: `url(${imageUrl})` }}></div>
+          <p className="profile-display-name mb-0 profile-details">{displayName + ' (' + artistType + ')'}</p>
+          <p className="profile-details mb-0">{firstName + ' ' + lastName}</p>
+          <p className="profile-details details-location">{city + ', ' + country}</p>
+        </div>
+      </div>
+    )
+  }
+}
